+++ src/asyncCall.test.ts
@@ -288,7 +288,7 @@ describe("useDebouncedAsyncCall", () => {
     expect(isWaiting).toBe(false);
   });
 
-  it("should correctly reset waiting state when leading = true and call is invoked only once", async () => {
+  it("should reset waiting state when leading = true and call is invoked only once", async () => {
     const { func, resolves } = createMockFunc();
     const t = renderHook(() =>
       useDebouncedAsyncCall({
@@ -323,72 +323,6 @@ describe("useDebouncedAsyncCall", () => {
     act(() => {
       jest.advanceTimersByTime(1000);
     });
-    expect(func).toHaveBeenCalledTimes(1);
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("FOO");
-    expect(isWaiting).toBe(false);
-  });
-
-  it("should skip function call if shouldCall returns false", async () => {
-    const shouldCall = jest.fn<boolean, [[string], [string]]>(
-      ([prev], [next]) => prev.toUpperCase() !== next.toUpperCase()
-    );
-    const { func, resolves } = createMockFunc();
-    const t = renderHook(() =>
-      useDebouncedAsyncCall({
-        func,
-        init: "",
-        wait: 1000,
-        shouldCall,
-      })
-    );
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).not.toHaveBeenCalled();
-    const [, call] = t.result.current;
-    let [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(false);
-
-    act(() => {
-      call("foo");
-    });
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).not.toHaveBeenCalled();
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(true);
-
-    act(() => {
-      jest.advanceTimersByTime(1000);
-    });
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(1);
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(true);
-
-    resolves[0]("FOO");
-    await t.waitForNextUpdate();
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(1);
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("FOO");
-    expect(isWaiting).toBe(false);
-
-    act(() => {
-      call("FOO");
-    });
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(1);
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("FOO");
-    expect(isWaiting).toBe(true);
-
-    act(() => {
-      jest.advanceTimersByTime(1000);
-    });
-    expect(shouldCall).toHaveBeenCalledTimes(1);
-    expect(shouldCall).toHaveBeenLastCalledWith(["foo"], ["FOO"]);
     expect(func).toHaveBeenCalledTimes(1);
     [res, , isWaiting] = t.result.current;
     expect(res).toBe("FOO");
@@ -559,81 +493,6 @@ describe("useDebouncedAsyncCall", () => {
     expect(func).toHaveBeenCalledTimes(2);
     [res, , isWaiting] = t.result.current;
     expect(res).toBe("BAR");
-    expect(isWaiting).toBe(false);
-  });
-
-  it("should reset the internal memory on rejection to ensure the function is called again even if shouldCall will return false", async () => {
-    const shouldCall = jest.fn<boolean, [[string], [string]]>(
-      ([prev], [next]) => prev.toUpperCase() !== next.toUpperCase()
-    );
-    const { func, resolves, rejects } = createMockFunc();
-    const t = renderHook(() =>
-      useDebouncedAsyncCall({
-        func,
-        init: "",
-        wait: 1000,
-        shouldCall,
-      })
-    );
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).not.toHaveBeenCalled();
-    const [, call] = t.result.current;
-    let [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(false);
-
-    act(() => {
-      call("foo");
-    });
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).not.toHaveBeenCalled();
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(true);
-
-    act(() => {
-      jest.advanceTimersByTime(1000);
-    });
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(1);
-    expect(func).toHaveBeenLastCalledWith("foo");
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(true);
-
-    rejects[0](new Error("test error"));
-    await t.waitForNextUpdate();
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(1);
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(false);
-
-    act(() => {
-      call("FOO");
-    });
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(1);
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(true);
-
-    act(() => {
-      jest.advanceTimersByTime(1000);
-    });
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(2);
-    expect(func).toHaveBeenLastCalledWith("FOO");
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("");
-    expect(isWaiting).toBe(true);
-
-    resolves[1]("FOO");
-    await t.waitForNextUpdate();
-    expect(shouldCall).not.toHaveBeenCalled();
-    expect(func).toHaveBeenCalledTimes(2);
-    [res, , isWaiting] = t.result.current;
-    expect(res).toBe("FOO");
     expect(isWaiting).toBe(false);
   });
 
@@ -852,89 +711,6 @@ describe("useDebouncedAsyncCall", () => {
       expect(res).toBe("");
       expect(isWaiting).toBe(false);
     });
-
-    it("should reset the internal memory to ensure the function is called again even if shouldCall will return false", async () => {
-      const shouldCall = jest.fn<boolean, [[string], [string]]>(
-        ([prev], [next]) => prev.toUpperCase() !== next.toUpperCase()
-      );
-      const { func, resolves } = createMockFunc();
-      const t = renderHook(() =>
-        useDebouncedAsyncCall({
-          func,
-          init: "",
-          wait: 1000,
-          shouldCall,
-        })
-      );
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).not.toHaveBeenCalled();
-      const [, call, , { cancel }] = t.result.current;
-      let [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(false);
-
-      act(() => {
-        call("foo");
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).not.toHaveBeenCalled();
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(true);
-
-      act(() => {
-        jest.advanceTimersByTime(1000);
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      expect(func).toHaveBeenLastCalledWith("foo");
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(true);
-
-      act(() => {
-        cancel();
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(false);
-
-      resolves[0]("FOO"); // should not update the state
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(false);
-
-      act(() => {
-        call("FOO");
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(true);
-
-      act(() => {
-        jest.advanceTimersByTime(1000);
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(2);
-      expect(func).toHaveBeenLastCalledWith("FOO");
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(true);
-
-      resolves[1]("FOO");
-      await t.waitForNextUpdate();
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(2);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("FOO");
-      expect(isWaiting).toBe(false);
-    });
   });
 
   describe("reset", () => {
@@ -1046,90 +822,6 @@ describe("useDebouncedAsyncCall", () => {
       expect(func).not.toHaveBeenCalled();
       [res, , isWaiting] = t.result.current;
       expect(res).toBe("RESET");
-      expect(isWaiting).toBe(false);
-    });
-
-    it("should reset the internal memory to ensure the function is called again even if shouldCall will return false", async () => {
-      const shouldCall = jest.fn<boolean, [[string], [string]]>(
-        ([prev], [next]) => prev.toUpperCase() !== next.toUpperCase()
-      );
-      const { func, resolves } = createMockFunc();
-      const t = renderHook(() =>
-        useDebouncedAsyncCall({
-          func,
-          init: "",
-          wait: 1000,
-          shouldCall,
-        })
-      );
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).not.toHaveBeenCalled();
-      const [, call, , { reset }] = t.result.current;
-      let [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(false);
-
-      act(() => {
-        call("foo");
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).not.toHaveBeenCalled();
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(true);
-
-      act(() => {
-        jest.advanceTimersByTime(1000);
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      expect(func).toHaveBeenLastCalledWith("foo");
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("");
-      expect(isWaiting).toBe(true);
-
-      resolves[0]("FOO");
-      await t.waitForNextUpdate();
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("FOO");
-      expect(isWaiting).toBe(false);
-
-      act(() => {
-        reset("RESET");
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("RESET");
-      expect(isWaiting).toBe(false);
-
-      act(() => {
-        call("FOO");
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(1);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("RESET");
-      expect(isWaiting).toBe(true);
-
-      act(() => {
-        jest.advanceTimersByTime(1000);
-      });
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(2);
-      expect(func).toHaveBeenLastCalledWith("FOO");
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("RESET");
-      expect(isWaiting).toBe(true);
-
-      resolves[1]("FOO");
-      await t.waitForNextUpdate();
-      expect(shouldCall).not.toHaveBeenCalled();
-      expect(func).toHaveBeenCalledTimes(2);
-      [res, , isWaiting] = t.result.current;
-      expect(res).toBe("FOO");
       expect(isWaiting).toBe(false);
     });
   });